Segregate Source chain events into separate mappers
